Update game if it doesnt have bestPlayers, ignore expansions and not owned games
