+++ routes/user.js
@@ -2,11 +2,9 @@ const mongoose = require('mongoose');
 const rp = require('request-promise');
 const parser = require('xml2json');
 
-const { getBestPlayerNumber, roundRating } = require('../lib/bbg-api-parse');
-const { shouldUpdateGame } = require('../lib/game-utils');
+const { saveGameFromBBG } = require('../lib/game-utils');
 
 const User = mongoose.model('users'); // Dont import models, access them like this via mongoose
-const Game = mongoose.model('games');
 
 module.exports = app => {
 	app.post('/api/add-bbg-username', (req, res) => {
@@ -23,43 +21,10 @@ module.exports = app => {
 
 		rp(options)
 			.then(body => {
-				body.forEach(bbgGame => {
-					Game.findOne({ _id: bbgGame.gameId })
-						.then(dbGame => {
-							if (shouldUpdateGame(dbGame, bbgGame)) {
-								const {
-									gameId,
-									name,
-									image,
-									maxPlayers,
-									minPlayers,
-									averageRating,
-								} = bbgGame;
-								const rating = roundRating(averageRating);
-								const game = {
-									_id: gameId,
-									name,
-									image,
-									maxPlayers,
-									minPlayers,
-									rating,
-								};
-								rp(`https://www.boardgamegeek.com/xmlapi/boardgame/${gameId}?&stats=1`)
-									.then(_body => {
-										const json = JSON.parse(parser.toJson(_body));
-										game.bestPlayers = getBestPlayerNumber(json);
-										(new Game(game)).save();
-									})
-									.catch(err => {
-										console.log('Error Fetching Rating');
-										console.log(err);
-										(new Game(game)).save();
-									});
-							}
-						});
-				});
+				body.forEach( bbgGame => saveGameFromBBG(bbgGame));
 			})
 			.catch(err => {
+				console.log('Error Fetching Game List');
 				console.log(err);
 			});
 	});
